Added ability to turn off Molcart support

+++ src/bin/hooks/sgc_audit_client.py
@@ -17,7 +17,7 @@ import cx_Oracle
 import MySQLdb as mdb
 
 class SGCAuditClient(AuditClient):
-    def __init__(self, hostname, port, username, password, projects, no_records, oracle_info, mysql_info):
+    def __init__(self, hostname, port, username, password, projects, no_records, oracle_info, mysql_info, has_molcart):
         self.bh = cx_Oracle.connect(oracle_info['username']+'/'+oracle_info['password']+'@'+oracle_info['tns_name'])
         self.mbh = mdb.connect(mysql_info['hostname'], mysql_info['username'], mysql_info['password'])
 
@@ -32,6 +32,7 @@ class SGCAuditClient(AuditClient):
         self.insert_count = 0
         self.delete_count = 0
         self.update_count = 0
+        self.has_molcart = has_molcart
         self.structure_updates = 0
 
         super(SGCAuditClient, self).__init__(hostname,  port, username, password, transaction_id, projects, no_records)
@@ -154,9 +155,10 @@ class SGCAuditClient(AuditClient):
     
         # We have to be sure to close the MySQL connection before calling our ICM script
         # because this Python script might lock the Molcart table
-        self.mbh.commit()
+        if self.has_molcart:
+            self.mbh.commit()
         
-        if self.structure_updates > 0:
+        if self.structure_updates > 0 and self.has_molcart:
             args = [
                 '/opt/local/icm/icm64',
                 os.path.dirname(os.path.abspath(__file__)) + '/molcart_insert.icm',
@@ -241,7 +243,7 @@ class SGCAuditClient(AuditClient):
                         update_blocks.append(self.fields[field] + ' = :' + self.fields[field])
                         update_values[':' + self.fields[field]] = item[field]
 
-                        if self.is_project_compound[project] and field == 'compound_id':
+                        if self.is_project_compound[project] and field == 'compound_id' and self.has_molcart:
                             self.mbh.cursor().execute('UPDATE GlobalCompounds.allOxford set SgcGlobalId=%s where ChemiRegPKEY=%s and ChemiRegPKEY is not null', (item[field],item['id']))
 
             if item['transaction_id'] > self.last_transaction_id:
@@ -298,7 +300,8 @@ class SGCAuditClient(AuditClient):
                 self.delete_from_molcart(item['id'])  
  
     def delete_from_molcart(self, id):
-        delete_from_molcart = self.mbh.cursor()
-        delete_from_molcart.execute('''
-            delete from GlobalCompounds.allOxford where ChemiRegPKEY = %s
-        ''',(id,))
+        if self.has_molcart:
+            delete_from_molcart = self.mbh.cursor()
+            delete_from_molcart.execute('''
+                delete from GlobalCompounds.allOxford where ChemiRegPKEY = %s
+            ''',(id,))
